flask-backend/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
boto3==1.14.56
botocore==1.17.56
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
flask_cors==3.0.10
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
numpy==1.19.2
pandas==1.1.2
loguru===0.5.3
protobuf==3.13.0
PyMySQL==0.10.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
urllib3==1.25.10
virtualenv==20.0.31
Werkzeug==1.0.1
zipp==3.1.0
